Create Selection Sort
In selection sort, we initially consider the first element as the minimum element. We then compare this element with the elements succeeding it in the array. If any of these succeeding elements are smaller than the element at the current minimum index, we update the minimum index to that element and continue. Finally, we swap the current element with the minimum element found during the iteration